lab3/requirements.txt
xarm-python-sdk
numpy<2
torch
robomimic